refactor(api-query-jpa): ActiveDeviceManager , ActiveMemberConnectionManager 구현
- MemberConnection 과 ActiveDevice 의 상태를 관리하는 클래스
